refactor: split into per-coin subpackages
Restructure the code so that coin-specific logic is separated from the
coin-agnostic core, in preparation for adding other coins (#52):

- rename eth_definitions/ -> definitions/
- definitions/common.py: keep only the coin-agnostic core (file format,
  payload encoding, shared helpers)
- definitions/serialize.py: serialization dispatch over all coins
- definitions/downloader.py: coin-agnostic downloading/caching
- definitions/download.py: slimmed down to pipeline orchestration
- definitions/ethereum/: ETH types, serialization, data sources
  (ethereum-lists, ERC-7730 registry, on-chain), builtin defs check
- definitions/solana/: SOL types, serialization, data sources

No behavioral change.
