+++ definitions/crypto.py
