+++ definitions/sign.py
