--- eth_definitions/download.py
@@ -1,984 +1,0 @@
-#!/usr/bin/env python3
-from __future__ import annotations
-
-import json
-import logging
-import re
-import sys
-import time
-from collections import Counter, defaultdict
-from pathlib import Path
-from typing import Any
-from urllib.parse import urlencode
-
-import click
-import requests
-from requests.adapters import HTTPAdapter
-from trezorlib import tools
-from urllib3.util.retry import Retry
-
-from .builtin_defs import check_builtin_defs
-from .check_definitions import check_definitions_list
-from .common import (
-    DEFINITIONS_PATH,
-    DISPLAY_FORMATS_LOG_PATH,
-    ChangeResolutionStrategy,
-    DefinitionsData,
-    ERC20DisplayFormat,
-    ERC20Token,
-    Network,
-    SolanaToken,
-    load_json_file,
-    make_metadata,
-    setup_logging,
-    store_definitions_data,
-)
-from .erc7730 import UnsupportedFeature, load_display_formats
-from .onchain import OnchainDecimalsResolver
-
-HERE = Path(__file__).parent
-ROOT_DIR = HERE.parent
-ETHEREUM_LISTS = ROOT_DIR / "ethereum-lists"
-ERC7730_REGISTRY = ROOT_DIR / "ethereum" / "clear-signing-erc7730-registry"
-
-TESTNET_WORDS = ("testnet", "devnet")
-
-CACHE_PATH = HERE / "definitions-cache.json"
-
-NETWORKS_PATH = ETHEREUM_LISTS / "chains" / "_data" / "chains"
-TOKENS_PATH = ETHEREUM_LISTS / "tokens" / "tokens"
-DISPLAY_FORMATS_PATH = ERC7730_REGISTRY / "registry"
-
-# Registry providers (top-level directory names) whose descriptors we emit.
-# The whole registry is still scanned so the log inventories what the other
-# providers would cost us, but only these feed the emitted records.
-ENABLED_PROVIDERS = frozenset(
-    {
-        "1inch",
-        "aave",
-        "consensus-specs",
-        "corestake",
-        "ethena",
-        "hyperliquid",
-        "kiln",
-        "lido",
-        "lifi",
-        "lombard",
-        "opencover",
-        "p2p",
-        "poap",
-        "quickswap",
-        "sei",
-        "starkgate",
-        "swell",
-        "tether",
-        "walletconnect",
-        "weth",
-        "yieldxyz",
-    }
-)
-
-
-class CacheableError(Exception):
-    def __init__(self, error_code: int):
-        self.error_code = error_code
-
-
-# ====== utils ======
-
-
-class CachedDict(dict[str, Any]):
-    """Generic cache object that caches to json."""
-
-    def __init__(self, cache_file: Path) -> None:
-        self.cache_file = cache_file
-        self.dirty = False
-        if not self.cache_file.exists():
-            self.cache_file.write_text("{}\n")
-        self.load()
-
-    def is_valid(self) -> bool:
-        return not self._is_empty() and not self._is_expired()
-
-    def _is_empty(self) -> bool:
-        return len(self) == 0
-
-    def _is_expired(self) -> bool:
-        mtime = self.cache_file.stat().st_mtime if self.cache_file.exists() else 0
-        time_diff = time.time() - mtime
-        return time_diff > 3600
-
-    def load(self) -> None:
-        self.clear()
-        self.update(json.loads(self.cache_file.read_text()))
-
-    def save(self, force: bool = False) -> None:
-        if not self.dirty and not force:
-            return
-        jsontext = json.dumps(self, sort_keys=True, indent=1)
-        self.cache_file.write_text(jsontext + "\n")
-        self.dirty = False
-
-    def __setitem__(self, key: str, value: Any) -> None:
-        super().__setitem__(key, value)
-        self.dirty = True
-
-
-class Downloader:
-    """Class that handles all the downloading and caching of Ethereum definitions."""
-
-    def __init__(
-        self, refresh: bool | None = None, sleep_duration: float = 0.0
-    ) -> None:
-        """
-        Args:
-            refresh: If True, force refresh of data. If False, use cached data. If None,
-            use cached data if available, otherwise force refresh.
-        """
-        self.cache = CachedDict(CACHE_PATH)
-        self.sleep_duration = sleep_duration
-        self.refresh = refresh
-        if refresh is None and not self.cache.is_valid():
-            self.refresh = True
-        self._init_requests_session()
-
-    def save_cache(self):
-        self.cache.save()
-
-    def _download_json(self, url: str, **url_params: Any) -> Any:
-        params = None
-        encoded_params = None
-        key = url
-
-        # convert params to lower-case strings (especially for boolean values
-        # because for CoinGecko API "True" != "true")
-        if url_params:
-            params = {key: str(value).lower() for key, value in url_params.items()}
-            encoded_params = urlencode(sorted(params.items()))
-            key += "?" + encoded_params
-
-        if self.refresh is False and key not in self.cache:
-            # refresh was explicitly disabled and key not found in cache
-            raise ValueError(f"Key {key} not found in cache")
-
-        if self.refresh is not True:
-            # refresh was not explicitly enabled, so use cached data if available
-            cached_result = self.cache.get(key)
-            if cached_result is not None:
-                if isinstance(cached_result, dict) and "error" in cached_result:
-                    raise CacheableError(cached_result["error"])
-                return cached_result
-
-        logging.info(f"Fetching data from {url}")
-
-        r = self.session.get(url, params=encoded_params, timeout=60)
-        if r.status_code == requests.codes.forbidden:
-            self.cache[key] = {"error": r.status_code}
-            raise CacheableError(r.status_code)
-        r.raise_for_status()
-        data = r.json()
-        self.cache[key] = data
-        if self.sleep_duration:
-            time.sleep(self.sleep_duration)
-        return data
-
-    def _init_requests_session(self) -> None:
-        self.session = requests.Session()
-        # As CoinGecko API will block us after ~30 requests for the whole minute,
-        # we need a way to retry the request multiple times.
-        retries = Retry(total=5, status_forcelist=[502, 503, 504])
-        self.session.mount("https://", HTTPAdapter(max_retries=retries))
-
-    def get_coingecko_asset_platforms(self) -> Any:
-        url = "https://api.coingecko.com/api/v3/asset_platforms"
-        return self._download_json(url)
-
-    def get_defillama_chains(self) -> Any:
-        url = "https://api.llama.fi/chains"
-        return self._download_json(url)
-
-    def get_coingecko_tokens_for_network(self, coingecko_network_id: str) -> list[Any]:
-        url = f"https://tokens.coingecko.com/{coingecko_network_id}/all.json"
-        try:
-            data = self._download_json(url)
-            return data.get("tokens", [])
-        except CacheableError:
-            # "Forbidden" is raised by Coingecko if no tokens are available under specified id
-            pass
-        except requests.exceptions.HTTPError as err:
-            raise err
-
-        return []
-
-    def get_coingecko_coins_list(self) -> Any:
-        url = "https://api.coingecko.com/api/v3/coins/list"
-        return self._download_json(url, include_platform=True)
-
-    def get_coingecko_top100(self) -> Any:
-        url = "https://api.coingecko.com/api/v3/coins/markets"
-        return self._download_json(
-            url,
-            vs_currency="usd",
-            order="market_cap_desc",
-            per_page=100,
-            page=1,
-            sparkline=False,
-        )
-
-
-def _get_testnet_status(*strings: str) -> bool:
-    for s in strings:
-        for testnet in TESTNET_WORDS:
-            if testnet in s.lower():
-                return True
-
-    return False
-
-
-# Overrides for networks whose ethereum-lists definition we do not want to
-# use. Keyed by chain_id.
-NETWORK_OVERRIDES: dict[int, Network] = {
-    # Chain id 999 is "Wanchain Testnet" in ethereum-lists, but HyperEVM
-    # (Hyperliquid) uses the same chain id and has far more usage, so we
-    # define it instead. HyperEVM is not present in ethereum-lists at all.
-    999: Network(
-        chain="hype",
-        chain_id=999,
-        is_testnet=False,
-        name="HyperEVM",
-        shortcut="HYPE",
-        slip44=60,
-    ),
-}
-
-
-def _load_ethereum_networks_from_repo() -> list[Network]:
-    """Load ethereum networks from submodule."""
-    networks: list[Network] = []
-    for chain in sorted(
-        NETWORKS_PATH.glob("eip155-*.json"),
-        key=lambda x: int(x.stem.replace("eip155-", "")),
-    ):
-        chain_data = load_json_file(chain)
-        shortcut = chain_data["nativeCurrency"]["symbol"]
-        name = chain_data["name"]
-        title = chain_data.get("title", "")
-        is_testnet = _get_testnet_status(name, title)
-        if is_testnet:
-            slip44 = 1
-        else:
-            slip44 = chain_data.get("slip44", 60)
-
-        if is_testnet and not shortcut.lower().startswith("t"):
-            shortcut = "t" + shortcut
-
-        # strip out bullcrap in network naming
-        if "mainnet" in name.lower():
-            name = re.sub(r" mainnet.*$", "", name, flags=re.IGNORECASE)
-
-        coin = NETWORK_OVERRIDES.get(chain_data["chainId"]) or Network(
-            chain=chain_data["shortName"],
-            chain_id=chain_data["chainId"],
-            is_testnet=is_testnet,
-            name=name,
-            shortcut=shortcut,
-            slip44=slip44,
-        )
-        networks.append(coin)
-
-    return networks
-
-
-def _build_token(
-    complex_token: dict[str, Any], chain_id: int, chain: str
-) -> ERC20Token | None:
-    # simple validation
-    if complex_token["address"][:2] != "0x" or int(complex_token["decimals"]) < 0:
-        return None
-    try:
-        bytes.fromhex(complex_token["address"][2:])
-    except ValueError:
-        return None
-
-    return ERC20Token(
-        address=str(complex_token["address"]).lower(),
-        chain=chain,
-        chain_id=chain_id,
-        decimals=complex_token["decimals"],
-        name=complex_token["name"],
-        shortcut=complex_token["symbol"],
-    )
-
-
-def _load_erc20_tokens_from_coingecko(
-    downloader: Downloader, networks: list[Network]
-) -> list[ERC20Token]:
-    tokens: list[ERC20Token] = []
-    for network in networks:
-        network_id = network.get("coingecko_network_id")
-        if network_id is None:
-            network_id = network.get("coingecko_id")
-        if network_id is None:
-            continue
-
-        all_tokens = downloader.get_coingecko_tokens_for_network(network_id)
-
-        for token in all_tokens:
-            t = _build_token(token, network["chain_id"], network["chain"])
-            if t is not None:
-                tokens.append(t)
-
-    return tokens
-
-
-def _load_erc20_tokens_from_repo(networks: list[Network]) -> list[ERC20Token]:
-    """Load ERC20 tokens from submodule."""
-    tokens: list[ERC20Token] = []
-    for network in networks:
-        chain_path = TOKENS_PATH / network["chain"]
-        for file in chain_path.glob("*.json"):
-            token = load_json_file(file)
-            t = _build_token(token, network["chain_id"], network["chain"])
-            if t is not None:
-                tokens.append(t)
-
-    return tokens
-
-
-def _force_networks_fields_sizes_t1(networks: list[Network]) -> None:
-    """Check sizes of embedded network fields for Trezor model 1 based on
-    "legacy/firmware/protob/messages-ethereum.options"."""
-    # EthereumNetworkInfo.name     max_size:256
-    # EthereumNetworkInfo.shortcut max_size:256
-    limit = 256
-    for network in networks:
-        # Cutting of what is over the limit
-        if len(network["name"]) > limit:
-            logging.info(f"Shortening name in {network}")
-            network["name"] = network["name"][:limit]
-        if len(network["shortcut"]) > limit:
-            logging.info(f"Shortening shortcut in {network}")
-            network["shortcut"] = network["shortcut"][:limit]
-
-
-def _force_tokens_fields_sizes_t1(tokens: list[ERC20Token]) -> None:
-    """Check sizes of embeded token fields for Trezor model 1 based on
-    "legacy/firmware/protob/messages-ethereum.options"."""
-    # EthereumTokenInfo.name    max_size:256
-    # EthereumTokenInfo.symbol  max_size:256 (here stored under "shortcut")
-    # EthereumTokenInfo.address max_size:20
-    limit = 256
-    address_bytes_len = 20
-
-    idxs_to_remove: list[int] = []
-    for idx, token in enumerate(tokens):
-        # Check address length (starts with 0x) and mark token for removal if invalid
-        try:
-            address_bytes = bytes.fromhex(token["address"][2:])
-            if len(address_bytes) != address_bytes_len:
-                raise AssertionError
-        except (ValueError, AssertionError):
-            logging.warning(
-                f"\nWARNING: invalid address length - not including {token}."
-            )
-            idxs_to_remove.append(idx)
-            continue
-
-        # Cutting of what is over the limit
-        if len(token["name"]) > limit:
-            logging.info(f"Shortening name in {token}")
-            token["name"] = token["name"][:limit]
-        if len(token["shortcut"]) > limit:
-            logging.info(f"Shortening shortcut in {token}")
-            token["shortcut"] = token["shortcut"][:limit]
-
-    # Remove tokens marked for removal
-    idxs_to_remove.sort(reverse=True)
-    for idx in idxs_to_remove:
-        tokens.pop(idx)
-
-
-def _build_solana_token(complex_token: dict[str, Any]) -> SolanaToken | None:
-    """Build a Solana token from jup.ag data."""
-    # simple validation
-    if not complex_token.get("address") or not complex_token.get("symbol"):
-        return None
-
-    try:
-        tools.b58decode(complex_token["address"])
-    except Exception as e:
-        logging.warning(f"Failed to decode Solana token: {e}")
-        return None
-
-    return {
-        "mint": complex_token["address"],
-        "name": complex_token["name"],
-        "shortcut": complex_token["symbol"].upper(),
-    }
-
-
-def _load_display_formats_from_repo(
-    networks: list[Network],
-) -> list[ERC20DisplayFormat]:
-    """Load ERC-7730 calldata display formats from the registry submodule.
-
-    Only `calldata-*.json` files are scanned (the proto's `func_sig` is
-    calldata-specific). `common-*.json` files reach us via the `includes`
-    merge; `eip712-*.json` files aren't representable in our schema.
-
-    Skips files under `tests/` subdirectories and records for chain_ids we
-    don't otherwise know about.
-
-    A display format that uses any feature we can't faithfully represent is
-    skipped whole (we never emit one with a field missing). Every such feature,
-    and every accepted-but-adjusted field, is collected and written to
-    `definitions-latest.log`. The whole registry is scanned for that inventory,
-    but only providers in `ENABLED_PROVIDERS` (noted in the log header) feed
-    the emitted records.
-
-    Deduplicates on `(chain_id, address, func_sig)`;
-    later files override earlier ones.
-    """
-
-    known_chain_ids = {n["chain_id"] for n in networks}
-    unsupported: list[tuple[str, str, str]] = []
-    adjustments: list[tuple[str, str, str]] = []
-    loaded: list[tuple[str, bool, list[ERC20DisplayFormat]]] = []
-
-    for path in sorted(DISPLAY_FORMATS_PATH.glob("*/calldata-*.json")):
-        if "tests" in path.parts:
-            continue
-
-        try:
-            # Scan every file so the log covers the whole registry, regardless
-            # of which providers are enabled below.
-            records = load_display_formats(
-                path, unsupported=unsupported, adjustments=adjustments
-            )
-        except UnsupportedFeature as e:
-            # File skipped — its features were already collected into `unsupported`.
-            logging.info(f"skipping {path.relative_to(ROOT_DIR)} — {e}")
-            continue
-        except Exception:
-            logging.warning(f"failed to parse {path.relative_to(ROOT_DIR)}")
-            raise
-
-        rel = str(path.relative_to(ROOT_DIR))
-        # `*/calldata-*.json` glob → the provider is the file's parent directory.
-        gated = path.parent.name in ENABLED_PROVIDERS
-        loaded.append((rel, gated, records))
-
-    display_formats, conflicts = _dedup_display_formats(loaded, known_chain_ids)
-    for key_str, overridden, kept in conflicts:
-        logging.warning(
-            f"display-format override: {kept} redefines {key_str} (was {overridden})"
-        )
-
-    if loaded or unsupported:
-        _write_display_formats_log(unsupported, conflicts, adjustments)
-    else:
-        # Nothing scanned at all — the registry submodule is most likely
-        # uninitialized (e.g. a shallow checkout). Surface the probable cause
-        # instead of writing a misleading "0 files" log.
-        logging.warning(
-            f"no ERC-7730 descriptors found under {DISPLAY_FORMATS_PATH} — is the "
-            "clear-signing-erc7730-registry submodule initialized?"
-        )
-    return display_formats
-
-
-def _dedup_display_formats(
-    loaded: list[tuple[str, bool, list[ERC20DisplayFormat]]],
-    known_chain_ids: set[int],
-) -> tuple[list[ERC20DisplayFormat], list[tuple[str, str, str]]]:
-    """Deduplicate display formats on `(chain_id, address, func_sig)`.
-
-    `loaded` is a list of `(source, gated, records)` in processing order. Only
-    `gated` records feed the emitted output (later files override earlier ones),
-    but conflicts are detected registry-wide: whenever any two files define the
-    same key with a *different* payload it's reported as an override conflict
-    `(key, overridden_source, kept_source)` — even for not-enabled providers.
-    Identical redefinitions are harmless duplicates and not reported.
-    """
-    dedup: dict[tuple[int, str, str], ERC20DisplayFormat] = {}
-    seen: dict[tuple[int, str, str], tuple[ERC20DisplayFormat, str]] = {}
-    conflicts: list[tuple[str, str, str]] = []
-
-    for source, gated, records in loaded:
-        for r in records:
-            if r["chain_id"] not in known_chain_ids:
-                continue
-            key = (r["chain_id"], r["address"], r["func_sig"])
-            prev = seen.get(key)
-            if prev is not None and prev[0] != r:
-                key_str = f"chain={key[0]} address={key[1]} selector={key[2]}"
-                conflicts.append((key_str, prev[1], source))
-            seen[key] = (r, source)
-            if gated:
-                dedup[key] = r
-
-    return list(dedup.values()), conflicts
-
-
-def _group_by_kind_and_file(
-    entries: list[tuple[str, str, str]],
-) -> tuple[Counter[str], dict[str, set[str]], dict[str, list[tuple[str, str]]]]:
-    """Group `(source, kind, detail)` entries for the two log views."""
-    by_kind_count: Counter[str] = Counter(kind for _, kind, _ in entries)
-    by_kind_files: dict[str, set[str]] = defaultdict(set)
-    by_file: dict[str, list[tuple[str, str]]] = defaultdict(list)
-    for source, kind, detail in entries:
-        by_kind_files[kind].add(source)
-        by_file[source].append((kind, detail))
-    return by_kind_count, by_kind_files, by_file
-
-
-def _grouped_section_lines(
-    entries: list[tuple[str, str, str]], kind_header: str
-) -> list[str]:
-    by_kind_count, by_kind_files, by_file = _group_by_kind_and_file(entries)
-    lines = ["", kind_header]
-    if not entries:
-        lines.append("(none)")
-    for kind in sorted(by_kind_count, key=lambda k: (-by_kind_count[k], k)):
-        lines.append(
-            f"{by_kind_count[kind]:5d}  {len(by_kind_files[kind]):4d} file(s)  {kind}"
-        )
-    lines += ["", "## By file"]
-    for source in sorted(by_file):
-        lines.append(source)
-        for kind, detail in sorted(by_file[source]):
-            lines.append(f"    {kind}: {detail}")
-        lines.append("")
-    return lines
-
-
-def _write_display_formats_log(
-    unsupported: list[tuple[str, str, str]],
-    conflicts: list[tuple[str, str, str]],
-    adjustments: list[tuple[str, str, str]],
-) -> None:
-    """Write the ERC-7730 processing log (`definitions-latest.log`).
-
-    Headed by the provider allowlist, then three independent sections: dropped
-    display formats grouped by unsupported feature, accepted-but-adjusted
-    fields (formatter overrides, ABI retypes, materialized constants), and
-    conflicting overrides (the same key redefined differently by multiple
-    files). All three cover the whole registry, not just enabled providers.
-    """
-    enabled = ", ".join(sorted(ENABLED_PROVIDERS)) if ENABLED_PROVIDERS else "(none)"
-    affected_files = {src for src, _, _ in unsupported}
-    lines = [
-        f"# Providers enabled: {enabled}",
-        "",
-        "# ERC-7730 unsupported features (display format dropped)",
-        f"# {len(affected_files)} file(s) affected, "
-        f"{len(unsupported)} feature occurrence(s)",
-    ]
-    lines += _grouped_section_lines(unsupported, "## By feature")
-
-    adjusted_files = {src for src, _, _ in adjustments}
-    lines += [
-        "",
-        "# Adjustments (field accepted with modification)",
-        f"# {len(adjusted_files)} file(s), {len(adjustments)} adjustment(s)",
-    ]
-    lines += _grouped_section_lines(adjustments, "## By kind")
-
-    lines += [
-        "",
-        "# Conflicting overrides (later file wins)",
-        f"# {len(conflicts)} key(s) redefined with a different payload",
-        "",
-    ]
-    if not conflicts:
-        lines.append("(none)")
-    for key_str, overridden, kept in sorted(conflicts):
-        lines.append(key_str)
-        lines.append(f"    kept:     {kept}")
-        lines.append(f"    overrode: {overridden}")
-        lines.append("")
-
-    DISPLAY_FORMATS_LOG_PATH.write_text("\n".join(lines).rstrip() + "\n")
-    logging.info(
-        f"wrote {DISPLAY_FORMATS_LOG_PATH.name} "
-        f"({len(affected_files)} file(s) with drops, {len(adjustments)} adjustment(s), "
-        f"{len(conflicts)} conflict(s))"
-    )
-
-
-def _load_solana_tokens_from_coingecko(downloader: Downloader) -> list[SolanaToken]:
-    """Load Solana tokens from coingecko API."""
-    tokens: list[SolanaToken] = []
-    all_tokens = downloader.get_coingecko_tokens_for_network("solana")
-    for token in all_tokens:
-        t = _build_solana_token(token)
-        if t is not None:
-            tokens.append(t)
-    return tokens
-
-
-def _update_display_formats_only(
-    networks: list[Network],
-    change_strategy: ChangeResolutionStrategy,
-    show_all: bool,
-    show_added: bool,
-) -> None:
-    """Refresh only the ERC-7730 display formats, leaving the networks, ERC-20 and
-    Solana token definitions untouched. No CoinGecko/DeFiLlama calls are made;
-    `networks` (loaded from the repo) only supplies the set of known chain ids.
-    """
-    if not DEFINITIONS_PATH.exists():
-        raise click.ClickException(
-            f"{DEFINITIONS_PATH} not found — run a full download first."
-        )
-
-    display_formats = _load_display_formats_from_repo(networks)
-
-    old_defs = load_json_file(DEFINITIONS_PATH)
-
-    def callback():
-        DEFINITIONS_PATH.write_text(json.dumps(old_defs, indent=2) + "\n")
-
-    check_definitions_list(
-        old_defs=old_defs.get("erc20_display_formats", []),
-        new_defs=display_formats,
-        change_strategy=change_strategy,
-        show_all=show_all,
-        show_added=show_added,
-        update_callback=callback,
-        main_keys=("chain_id", "address", "func_sig"),
-        def_type="DISPLAY_FORMAT",
-        # A wrong display format is worse than a missing one: no tombstones —
-        # formats the parser stops emitting are removed and stop being signed.
-        only_mark_as_deleted=False,
-    )
-
-    display_formats.sort(key=lambda x: (x["chain_id"], x["address"], x["func_sig"]))
-
-    # Reuse the existing networks/tokens/Solana definitions verbatim.
-    definitions_data = DefinitionsData(
-        networks=old_defs["networks"],
-        erc20_tokens=old_defs["erc20_tokens"],
-        solana_tokens=old_defs["solana_tokens"],
-        erc20_display_formats=display_formats,
-    )
-
-    metadata = make_metadata(definitions_data)
-    store_definitions_data(metadata, definitions_data)
-
-
-@click.command()
-@click.option(
-    "-r/-R",
-    "--refresh/--no-refresh",
-    default=None,
-    help="Force refresh or no-refresh data. By default tries to load cached data.",
-)
-@click.option(
-    "-i",
-    "--interactive",
-    is_flag=True,
-    help="Ask about every change in symbols/decimals.",
-)
-@click.option(
-    "--really-apply-all-renames-without-confirmation",
-    "force_changes",
-    is_flag=True,
-    help="Changes to symbols/decimals in definitions should be accepted.",
-)
-@click.option(
-    "-s",
-    "--show-all",
-    is_flag=True,
-    help="Show the differences of all definitions. By default only changes to top 100 definitions (by Coingecko market cap ranking) are shown.",
-)
-@click.option(
-    "-a",
-    "--show-added",
-    is_flag=True,
-    help="Show newly added definitions.",
-)
-@click.option(
-    "-c",
-    "--check-builtin",
-    is_flag=True,
-    help="Compares results with Trezor builtin definitions.",
-)
-@click.option(
-    "--sleep-duration",
-    type=float,
-    default=0,
-    help="Amount of seconds to sleep after each download.",
-)
-@click.option("-v", "--verbose", is_flag=True, help="Display more info")
-@click.option(
-    "--trace-address",
-    default=None,
-    help="Trace which source provided token info for this contract address, then exit.",
-)
-@click.option(
-    "--no-onchain-decimals",
-    is_flag=True,
-    help="Disable verifying conflicting token decimals against the contract on-chain.",
-)
-@click.option(
-    "--erc7730-only",
-    is_flag=True,
-    help="Refresh only the ERC-7730 display formats from the registry submodule, "
-    "reusing the existing networks/tokens/Solana definitions. Skips all CoinGecko "
-    "and DeFiLlama downloads.",
-)
-def download(
-    refresh: bool | None,
-    interactive: bool,
-    force_changes: bool,
-    show_all: bool,
-    show_added: bool,
-    check_builtin: bool,
-    verbose: bool,
-    sleep_duration: float,
-    trace_address: str | None,
-    no_onchain_decimals: bool,
-    erc7730_only: bool,
-) -> None:
-    """Download and prepare token definitions."""
-    setup_logging(verbose)
-
-    # validating change resolution strategy - max one of the options can be used
-    change_strategy = ChangeResolutionStrategy.from_args(
-        interactive=interactive,
-        force_accept=force_changes,
-    )
-
-    networks = _load_ethereum_networks_from_repo()
-
-    if erc7730_only:
-        _update_display_formats_only(
-            networks, change_strategy, show_all, show_added
-        )
-        return
-
-    # init Ethereum definitions downloader
-    downloader = Downloader(refresh, sleep_duration)
-
-    # coingecko API
-    cg_platforms_json = downloader.get_coingecko_asset_platforms()
-    cg_platforms: dict[int, tuple[str, str]] = {}
-    for chain in cg_platforms_json:
-        # We want only information about chains, that have both chain id and coingecko id,
-        # otherwise we could not link local and coingecko networks.
-        if (
-            chain["chain_identifier"] is not None
-            and chain["id"] is not None
-            and chain["native_coin_id"] is not None
-        ):
-            cg_platforms[chain["chain_identifier"]] = (
-                chain["id"],
-                chain["native_coin_id"],
-            )
-
-    # defillama API
-    dl_chains_json = downloader.get_defillama_chains()
-    dl_chains: dict[int, str] = {}
-    for chain in dl_chains_json:
-        # We want only information about chains, that have both chain id and coingecko id,
-        # otherwise we could not link local and coingecko networks.
-        if chain["chainId"] is not None and chain["gecko_id"] is not None:
-            dl_chains[chain["chainId"]] = chain["gecko_id"]
-
-    # We will try to get as many "coingecko_id"s as possible to be able to use them afterwards
-    # to load tokens from coingecko. We won't use coingecko networks, because we don't know which
-    # ones are EVM based.
-    network_to_cid: dict[str, int] = {}
-    native_coin_to_network: dict[str, Network] = {}
-    for network in networks:
-        # Assign coingecko_id if possible and not there already
-        chain_id = network["chain_id"]
-        if network.get("coingecko_id") is None:
-            # from coingecko via chain_id
-            if chain_id in cg_platforms:
-                network_id, cg_id = cg_platforms[chain_id]
-                network["coingecko_id"] = cg_id
-                network["coingecko_network_id"] = network_id
-                network_to_cid[network_id] = chain_id
-                native_coin_to_network[cg_id] = network
-            # from defillama via chain_id
-            elif chain_id in dl_chains:
-                network["coingecko_network_id"] = dl_chains[chain_id]
-
-        # if we found "coingecko_id" add it to the map - used later to map tokens with coingecko ids
-        if (network_id := network.get("coingecko_network_id")) is not None:
-            network_to_cid[network_id] = chain_id
-
-    # get tokens
-    cg_tokens = _load_erc20_tokens_from_coingecko(downloader, networks)
-    repo_tokens = _load_erc20_tokens_from_repo(networks)
-    solana_tokens = _load_solana_tokens_from_coingecko(downloader)
-
-    # get ERC-7730 display formats from the registry submodule
-    display_formats = _load_display_formats_from_repo(networks)
-
-    # get data used in further processing now to be able to save cache before we do any
-    # token collision process and others
-    # get CoinGecko coin list
-    cg_coin_list = downloader.get_coingecko_coins_list()
-    # get top 100 coins
-    cg_top100 = downloader.get_coingecko_top100()
-    # save cache
-    downloader.save_cache()
-
-    # merge tokens - CoinGecko have precedence, so starting with Ethereum repo first
-    token_deduplicator: dict[tuple[int, str], ERC20Token] = {}
-    for token in repo_tokens + cg_tokens:
-        token_deduplicator[(token["chain_id"], token["address"])] = token
-    erc20_tokens = list(token_deduplicator.values())
-
-    if trace_address:
-        addr = trace_address.lower()
-        print(f"\n=== TRACING ADDRESS: {addr} ===")
-
-        repo_match = next((t for t in repo_tokens if t["address"] == addr), None)
-        cg_match = next((t for t in cg_tokens if t["address"] == addr), None)
-
-        if repo_match:
-            for network in networks:
-                if network["chain_id"] == repo_match["chain_id"]:
-                    chain_path = TOKENS_PATH / network["chain"]
-                    for f in chain_path.glob("*.json"):
-                        data = load_json_file(f)
-                        if data.get("address", "").lower() == addr:
-                            print(f"  [REPO] File: {f}")
-                            print(f"  [REPO] Data: {repo_match}")
-                            break
-                    break
-
-        if cg_match:
-            for network in networks:
-                if network["chain_id"] == cg_match["chain_id"]:
-                    cg_network_id = network.get("coingecko_network_id") or network.get(
-                        "coingecko_id"
-                    )
-                    print(f"  [COINGECKO] Network ID: {cg_network_id}")
-                    print(
-                        f"  [COINGECKO] API URL: https://tokens.coingecko.com/{cg_network_id}/all.json"
-                    )
-                    print(f"  [COINGECKO] Data: {cg_match}")
-                    break
-
-        if not repo_match and not cg_match:
-            print("  NOT FOUND in any source.")
-        else:
-            winner_source = "COINGECKO" if cg_match is not None else "REPO"
-            print(
-                f"\n  >>> FINAL SOURCE: {winner_source} (CoinGecko overrides repo when both present)"
-            )
-
-        print("=== END TRACE ===\n")
-        sys.exit(0)
-
-    # remove items with empty symbol
-    networks = [n for n in networks if n["shortcut"]]
-    erc20_tokens = [t for t in erc20_tokens if t["shortcut"]]
-
-    # Enforce the maximum field sizes
-    _force_networks_fields_sizes_t1(networks)
-    _force_tokens_fields_sizes_t1(erc20_tokens)
-
-    # map coingecko ids to tokens
-    # NOTE: changes the `tokens` in place!
-    tokens_by_chain_id_and_address = {
-        (t["chain_id"], t["address"]): t for t in erc20_tokens
-    }
-    solana_tokens_by_mint = {(t["mint"]): t for t in solana_tokens}
-    for cg_coin in cg_coin_list:
-        for platform_name, address in cg_coin.get("platforms", {}).items():
-            key = (network_to_cid.get(platform_name), address)
-            if key in tokens_by_chain_id_and_address:
-                tokens_by_chain_id_and_address[key]["coingecko_id"] = cg_coin["id"]
-            if platform_name == "solana" and address in solana_tokens_by_mint:
-                solana_tokens_by_mint[address]["coingecko_id"] = cg_coin["id"]
-        # enrich networks by symbols known from coingecko
-        if (network := native_coin_to_network.get(cg_coin["id"])) is not None:
-            network["name"] = cg_coin["name"]
-            network["shortcut"] = cg_coin["symbol"].upper()
-
-    # get top 100 ids
-    cg_top100_ids = {d["id"]: d for d in cg_top100}
-
-    for item in networks + erc20_tokens + solana_tokens:
-        if (id := item.get("coingecko_id")) in cg_top100_ids:
-            item["coingecko_rank"] = cg_top100_ids[id]["market_cap_rank"]
-
-    if DEFINITIONS_PATH.exists():
-        old_defs = load_json_file(DEFINITIONS_PATH)
-
-        def callback():
-            DEFINITIONS_PATH.write_text(json.dumps(old_defs, indent=2) + "\n")
-
-        decimals_resolver = None if no_onchain_decimals else OnchainDecimalsResolver()
-
-        # check networks and tokens
-        check_definitions_list(
-            old_defs=old_defs["networks"],
-            new_defs=networks,
-            change_strategy=change_strategy,
-            show_all=show_all,
-            show_added=show_added,
-            update_callback=callback,
-        )
-        check_definitions_list(
-            old_defs=old_defs["erc20_tokens"],
-            new_defs=erc20_tokens,
-            change_strategy=change_strategy,
-            show_all=show_all,
-            show_added=show_added,
-            update_callback=callback,
-            decimals_resolver=decimals_resolver,
-        )
-        check_definitions_list(
-            old_defs=old_defs["solana_tokens"],
-            new_defs=solana_tokens,
-            change_strategy=change_strategy,
-            show_all=show_all,
-            show_added=show_added,
-            update_callback=callback,
-        )
-        check_definitions_list(
-            old_defs=old_defs.get("erc20_display_formats", []),
-            new_defs=display_formats,
-            change_strategy=change_strategy,
-            show_all=show_all,
-            show_added=show_added,
-            update_callback=callback,
-            main_keys=("chain_id", "address", "func_sig"),
-            def_type="DISPLAY_FORMAT",
-            # A wrong display format is worse than a missing one: no tombstones —
-            # formats the parser stops emitting are removed and stop being signed.
-            only_mark_as_deleted=False,
-        )
-
-    if check_builtin:
-        # check built-in definitions against generated ones
-        if not check_builtin_defs(networks, erc20_tokens):
-            logging.warning(
-                "\nWARNING: Built-in definitions differ from the generated ones."
-            )
-
-    # sort networks and tokens
-    networks.sort(key=lambda x: x["chain_id"])
-    erc20_tokens.sort(key=lambda x: (x["chain_id"], x["address"]))
-    solana_tokens.sort(key=lambda x: x["mint"])
-    display_formats.sort(key=lambda x: (x["chain_id"], x["address"], x["func_sig"]))
-
-    # create definitions data
-    definitions_data = DefinitionsData(
-        networks=networks,
-        erc20_tokens=erc20_tokens,
-        solana_tokens=solana_tokens,
-        erc20_display_formats=display_formats,
-    )
-
-    # save results
-    metadata = make_metadata(definitions_data)
-    store_definitions_data(metadata, definitions_data)
